Refactor: Adjust the location of the StatusService interface
